requirements.txt
streamlit
pandas
firebase-admin
google-cloud-firestore
google-auth
